refactor: migrate agent-router package to agents directory and update imports
- Moved the agent-router package to the agents directory, updating its structure and dependencies accordingly.
- Adjusted import paths in CLI commands to reference the new location of the agent-router.
- Removed the old agent-router package files, including tests and configuration, to streamline the project structure.
