+++ apps/cli/src/commands/agents-command.ts
@@ -1,5 +1,5 @@
 import { Args, Command, Options } from "@effect/cli";
-import { refreshAgentCache } from "@open-composer/agent-router";
+import { refreshAgentCache } from "agents/agent-router/src/index.js";
 import { type AgentCache, updateAgentCache } from "@open-composer/cache";
 import * as Effect from "effect/Effect";
 import * as Option from "effect/Option";
